CEDS-2688 Add .jvmopts file with new jvm options for sbt
After upgrading Scala version to 2.12.12, the compilation needs more
memory than before. This is probably caused by a large amount of implicit
values to resolve.
The easiest way to solve the memory problem is to assign more max
memory to the JVM with -Xmx setting. By using .jvmopts file SBT is
now overriding system JVM settings when starting a console in this
directory.