[CEDS-2295] Redesign file upload receipts page
As part of public beta readiness the content of the page was requested
to be altered by the UX team.

Main changes are adding a green panel that includes the MRN and links
to the NCH and 'print this page'.

This turned out to be a larger change than first envisaged as to
implement the green panel the new GDS frontend had to be incorporated
into the code base.

Due to time constraints upgrading all the pages to the new frontend
was not possible.