CEDS-2951 Clear cache at end of journey
Currently the user cache is cleared when the
user clicks on the 'start' button of the first
page in the journey.

Due to upcoming changes that means some users
won't see this first page (when they are linked
to SFUS directly from Exports with a passed MRN)
the clearing cache step is being moved to the end
of the user journey (the success or error page).