BAU Decrease number of Mongo connections in tests
Unit tests make use of Injector trait which creates new application
every time the trait is initialized. This results in hundreds of
connections being created every time all tests are run.

Application instance that was created in Injector trait is not possible
to be overridden so there should be no downsides of having a single
instance for all tests.