[CEDS-2946] Fix broken javascript upload form validation
While testing the new validation message text I noticed that
the javascript upload form field validation was not working.

I suspect it got broken when we upgraded SFUS to use the GDS
components.

To fix this I have had to add JQuery into the assets folder
and created a new flag for the main template to include the
JQuery scripts only for the page that requires it.