[CEDS-2946] Fix highlighting of form field error link
When the user fails to select an option in a radio
button form field, the error box displayed at the top
of the page should contain a link that when clicked
takes the user to that particular form field.

The two new choice pages on SFUS contained this link
but it was not taking them to the form field (was not
highlighting the first radio option) due to the anchor
reference being wrong. This is now fixed